requirements.txt
langchain==1.2.15
langchain-core==1.2.29
langchain-openai==1.1.13
langchain-community==0.4.1
langchain-text-splitters==1.1.1
chromadb==1.5.7
gradio==6.12.0
wikipedia==1.4.0
duckduckgo-search==8.1.1
python-dotenv==1.2.2
pypdf==6.10.2
arxiv==3.0.0
biopython==1.87
tiktoken==0.12.0
